Added Separate Login Page
- This is the first page the user sees
- Have not added all the functionality of the Login Pages

Missing functionality in login:
- Sign up and login button
- Allowing user to put a username and password
- Storing the username and password in local database
- After authentication and collection of all relevant data, redirecting the user to the application
