+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,12 @@
         tools:targetApi="31">
         <activity
             android:name=".ui.login.LoginActivity"
-            android:exported="false" />
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <activity
             android:name=".MainActivity"
             android:exported="true"
@@ -35,6 +40,7 @@
                     android:host="auth"
                     android:scheme="com.example.spotify2340" />
             </intent-filter>
+
             <!-- <intent-filter> -->
             <!-- <action android:name="android.intent.action.VIEW" /> -->
             <!-- <category android:name="android.intent.category.DEFAULT" /> -->
