+++ app/src/main/res/layout/activity_login.xml
@@ -13,7 +13,21 @@
         android:layout_height="wrap_content"
         android:background="@color/spotify_green"
         android:backgroundTint="@color/spotify_green"
-        android:text="Login"
+        android:text="Connect Account to Spotify"
+        android:textColor="@color/white"
+        app:cornerRadius="20pt"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+        android:id="@+id/connect_spotify_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:background="@color/spotify_green"
+        android:backgroundTint="@color/spotify_green"
+        android:text="Connect Account to Spotify"
         android:textColor="@color/white"
         app:cornerRadius="20pt"
         app:layout_constraintBottom_toBottomOf="parent"
