+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -3,7 +3,12 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/nav_account">
+    app:startDestination="@+id/nav_login">
+
+    <activity
+        android:id="@+id/nav_login"
+        android:name="com.example.spotify2340.ui.login.LoginActivity"
+        android:label="Login"/>
 
     <fragment
         android:id="@+id/nav_account"
